test(test): add a cargo-fuzz harness for the JSON validator
The validator runs over attacker-controlled bytes via `json validate` and
`sjq --validate`, so a panic is a denial of service — not hypothetical,
since #151 was a reachable stack-overflow abort. Nothing in the repo
searched for the next one; the property tests sample a space the fuzzer
can explore.

Three targets: no-panic (the one that justifies the harness, and the only
check that would catch memory unsafety under ASAN if the validator ever
grows an unsafe fast path), validity agreement with serde_json, and the
Position-reconstructible-from-offset invariant.

Isolated as its own workspace with its own lockfile, mirroring
bench-compare/: cargo-fuzz needs nightly for -Zsanitizer=address while
the main crate promises rust-version 1.73, so this keeps libfuzzer-sys
and arbitrary out of the main dependency graph entirely. Verified
`cargo deny check all` still passes and does not see the fuzz crate.

To stop the oracle drifting between the fuzzer and the property tests —
where a stale divergence classifier would silently excuse a real bug —
the shared logic moves to tests/common/json_oracle.rs and both consume it
by #[path] include. It lives in a subdirectory because cargo
auto-discovers tests/*.rs as test binaries but not files below them.

The seed corpus is generated by fuzz/seed-corpus.sh rather than
committed: every seed duplicates something already tracked (the 318
vendored JSONTestSuite cases, the real-workload bench seed, the proptest
regressions), so committing them again would leave two copies to sync.
libFuzzer requires every corpus directory named on the command line to
exist already, and cargo-fuzz only auto-creates the one it is passed
directly, so the script also creates each target's fuzz/corpus/<target>
directory up front — otherwise the two-directory invocation ("write
here, read seeds from there") fails on a fresh checkout for every target
that has never been run the default way.

Targets are compile-verified on nightly. Per the README they stay off the
pull-request path — a 60-second CI fuzz run finds nothing and grants
false confidence, which is the failure mode note_simd_skip exists to
prevent.
