+++ fuzz/Cargo.toml
@@ -1,0 +1,50 @@
+# Fuzzing harness for the JSON validator.
+#
+# Deliberately a SEPARATE workspace with its own lockfile, mirroring
+# bench-compare/. cargo-fuzz requires nightly (-Zsanitizer=address), while the
+# main crate promises rust-version 1.73; isolating the harness keeps
+# libfuzzer-sys and arbitrary out of the main dependency graph entirely, so
+# neither the MSRV promise nor `cargo deny check all` is affected.
+#
+# Usage (see fuzz/README.md):
+#   cargo +nightly fuzz run validate_never_panics
+[workspace]
+members = ["."]
+
+[package]
+name = "succinctly-fuzz"
+version = "0.0.0"
+publish = false
+edition = "2021"
+
+[package.metadata]
+cargo-fuzz = true
+
+[dependencies]
+libfuzzer-sys = "0.4"
+serde_json = "1.0"
+
+[dependencies.succinctly]
+path = ".."
+
+# Fuzz targets are binaries, not tests or benches.
+[[bin]]
+name = "validate_never_panics"
+path = "fuzz_targets/validate_never_panics.rs"
+test = false
+doc = false
+bench = false
+
+[[bin]]
+name = "validate_vs_serde_json"
+path = "fuzz_targets/validate_vs_serde_json.rs"
+test = false
+doc = false
+bench = false
+
+[[bin]]
+name = "validate_position_invariant"
+path = "fuzz_targets/validate_position_invariant.rs"
+test = false
+doc = false
+bench = false
